added tests for S3
