+++ requirements/test-requirements.txt
@@ -1,5 +1,6 @@
 pytest==7.1.1
 pytest-cov==3.0.0
 flake8==4.0.1
+moto==3.1.16
 piccolo[postgres,sqlite]>=0.30.0
 requests==2.27.1
